refactor(beacon): consolidate beacon client into beacon-client crate (#88)
Merge rvc/src/beacon features into beacon-client crate to eliminate code
duplication and establish a single source of truth for beacon client
functionality.

Changes:
- Add specialized endpoint methods to beacon-client:
  - get_attester_duties(epoch, validator_indices)
  - get_attestation_data(slot, committee_index)
  - submit_attestation(attestations) with partial failure handling
- Add missing types to beacon-client:
  - AttesterDutiesResponse and AttestationDataResponse type aliases
  - IndexedAttestationError struct
  - SubmitAttestationResult enum with is_success() and failures()
- Migrate comprehensive tests from rvc/src/beacon to beacon-client
- Update all rvc imports to use beacon_client crate
- Remove rvc/src/beacon module entirely

Closes #88
