+++ crates/beacon-client/src/lib.rs
@@ -9,6 +9,7 @@ mod types;
 pub use client::{BeaconClient, BeaconClientConfig};
 pub use error::BeaconError;
 pub use types::{
-    Attestation, AttestationData, AttesterDuty, BeaconBlockHeader, Checkpoint, DataResponse,
-    DependentRootResponse,
+    Attestation, AttestationData, AttestationDataResponse, AttesterDutiesResponse, AttesterDuty,
+    BeaconBlockHeader, Checkpoint, DataResponse, DependentRootResponse, IndexedAttestationError,
+    SubmitAttestationResult,
 };
