--- crates/rvc/src/beacon/error.rs
@@ -1,54 +1,0 @@
-use thiserror::Error;
-
-#[derive(Error, Debug)]
-pub enum BeaconError {
-    #[error("HTTP request failed: {0}")]
-    HttpError(String),
-
-    #[error("Failed to parse response: {0}")]
-    ParseError(String),
-
-    #[error("Beacon node returned error: {status} - {message}")]
-    ApiError { status: u16, message: String },
-
-    #[error("Request timeout")]
-    Timeout,
-
-    #[error("Invalid endpoint URL: {0}")]
-    InvalidUrl(String),
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_http_error_display() {
-        let err = BeaconError::HttpError("connection refused".to_string());
-        assert_eq!(err.to_string(), "HTTP request failed: connection refused");
-    }
-
-    #[test]
-    fn test_parse_error_display() {
-        let err = BeaconError::ParseError("invalid JSON".to_string());
-        assert_eq!(err.to_string(), "Failed to parse response: invalid JSON");
-    }
-
-    #[test]
-    fn test_api_error_display() {
-        let err = BeaconError::ApiError { status: 404, message: "Validator not found".to_string() };
-        assert_eq!(err.to_string(), "Beacon node returned error: 404 - Validator not found");
-    }
-
-    #[test]
-    fn test_timeout_error_display() {
-        let err = BeaconError::Timeout;
-        assert_eq!(err.to_string(), "Request timeout");
-    }
-
-    #[test]
-    fn test_invalid_url_error_display() {
-        let err = BeaconError::InvalidUrl("not a valid url".to_string());
-        assert_eq!(err.to_string(), "Invalid endpoint URL: not a valid url");
-    }
-}
